Conflicts:
	MaikelGerard/QL/GUI/DrawGUI.py
	MaikelGerard/QL/GUI/FormGUI.py
	MaikelGerard/StartQuestionnaire.py

+++ OffByOne/src/OffByOne.LanguageCore/Checker/BaseTypeCheckerVisitor.cs
@@ -21,7 +21,7 @@
 
         public ValueType Visit(DecimalLiteral literal, VisitorContext context)
         {
-            return TypeConstants.FloatType;
+            return TypeConstants.DecimalType;
         }
 
         public ValueType Visit(BooleanLiteral literal, VisitorContext context)
